BackendV1/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
openai
torch
ffmpeg-python
whisper
librosa
soundfile
pytesseract
opencv-python
ultralytics
